+ Corregidos los resultados esperados en testIntervalosIguales.

+++ tddTaller/src/test/java/org/colswe/junitlab/logica/ControlFestivoContarHorasTest.java
@@ -414,7 +414,7 @@ public class ControlFestivoContarHorasTest extends TestCase {
 		listIntervalos1.add(intervalo1);
 		listIntervalos1.add(intervalo2);
 		
-		assertEquals(new Integer(136), cf.contarHoras(listIntervalos1));
+		assertEquals(new Integer(240), cf.contarHoras(listIntervalos1));
 		
 		Intervalo intervalo3 = new Intervalo(fecha2, fecha3);
 		Intervalo intervalo4 = intervalo3;
@@ -423,7 +423,7 @@ public class ControlFestivoContarHorasTest extends TestCase {
 		listIntervalos2.add(intervalo3);
 		listIntervalos2.add(intervalo4);
 		
-		assertEquals(new Integer(120), cf.contarHoras(listIntervalos2));
+		assertEquals(new Integer(240), cf.contarHoras(listIntervalos2));
 
 	}
